Refine healthcheck wait strategy selection
