+++ packages/testcontainers/src/docker-compose-environment/docker-compose-environment.ts
@@ -1,24 +1,13 @@
-import { ContainerInfo, ContainerInspectInfo } from "dockerode";
+import { ContainerInfo } from "dockerode";
 import { containerLog, log, RandomUuid, Uuid } from "../common";
-import {
-  ComposeOptions,
-  ContainerRuntimeClient,
-  getContainerRuntimeClient,
-  ImageName,
-  parseComposeContainerName,
-} from "../container-runtime";
+import { ComposeOptions, getContainerRuntimeClient, parseComposeContainerName } from "../container-runtime";
 import { StartedGenericContainer } from "../generic-container/started-generic-container";
 import { getReaper } from "../reaper/reaper";
 import { Environment } from "../types";
 import { BoundPorts } from "../utils/bound-ports";
 import { mapInspectResult } from "../utils/map-inspect-result";
 import { ImagePullPolicy, PullPolicy } from "../utils/pull-policy";
-import {
-  hasDisabledHealthCheckConfig,
-  hasHealthCheckConfig,
-  hasHealthCheckStatus,
-} from "../wait-strategies/utils/health-check";
-import { Wait } from "../wait-strategies/wait";
+import { selectWaitStrategy } from "../wait-strategies/utils/wait-strategy-selector";
 import { waitForContainer } from "../wait-strategies/wait-for-container";
 import { WaitStrategy } from "../wait-strategies/wait-strategy";
 import { StartedDockerComposeEnvironment } from "./started-docker-compose-environment";
@@ -185,7 +174,11 @@ export class DockerComposeEnvironment {
           const inspectResult = await client.container.inspect(container);
           const mappedInspectResult = mapInspectResult(inspectResult);
           const boundPorts = BoundPorts.fromInspectResult(client.info.containerRuntime.hostIps, mappedInspectResult);
-          const waitStrategy = await this.selectWaitStrategy(client, containerName, inspectResult);
+          const waitStrategy = await selectWaitStrategy({
+            client,
+            inspectResult,
+            waitStrategy: this.waitStrategy[containerName] ?? this.defaultWaitStrategy,
+          });
           if (this.startupTimeoutMs !== undefined) {
             waitStrategy.withStartupTimeout(this.startupTimeoutMs);
           }
@@ -230,52 +223,6 @@ export class DockerComposeEnvironment {
       composeOptions,
       environment: this.environment,
     });
-  }
-
-  private async selectWaitStrategy(
-    client: ContainerRuntimeClient,
-    containerName: string,
-    inspectResult: ContainerInspectInfo
-  ): Promise<WaitStrategy> {
-    const containerWaitStrategy = this.waitStrategy[containerName]
-      ? this.waitStrategy[containerName]
-      : this.defaultWaitStrategy;
-    if (containerWaitStrategy) return containerWaitStrategy;
-    if (hasDisabledHealthCheckConfig(inspectResult)) {
-      return Wait.forListeningPorts();
-    }
-    if (hasHealthCheckConfig(inspectResult) || hasHealthCheckStatus(inspectResult)) {
-      return Wait.forHealthCheck();
-    }
-    if (await this.imageHasHealthCheck(client, inspectResult)) {
-      return Wait.forHealthCheck();
-    }
-    return Wait.forListeningPorts();
-  }
-
-  private async imageHasHealthCheck(
-    client: ContainerRuntimeClient,
-    inspectResult: ContainerInspectInfo
-  ): Promise<boolean> {
-    const imageNames = Array.from(
-      new Set(
-        [inspectResult.Config.Image, inspectResult.Image].filter(
-          (imageName): imageName is string => imageName !== undefined && imageName !== ""
-        )
-      )
-    );
-
-    for (const imageName of imageNames) {
-      try {
-        if (hasHealthCheckConfig(await client.image.inspect(ImageName.fromString(imageName)))) {
-          return true;
-        }
-      } catch (err) {
-        log.debug(`Failed to inspect image "${imageName}" for health check config: ${err}`);
-      }
-    }
-
-    return false;
   }
 
   private warnForUnusedWaitStrategies(startedContainerNames: Set<string>): void {
